Last commit for the night on this I think, but we don't need to call random twice. Before this commit we actually had mysql doing RANDOM, then doing return @data[int(rand(@data))] on an array of length _one_. I've changed this to return $data[0]; for the time being.
The best way to go from here is to keep a hash with matching channels ->
users with the value of the hash being the last message printed out by
the bot. re-request a quote if the hash and the user matches up. If you
get the same quote 3 times, give up and print it again. (or count the
number of entries in the database for that user, if its 1... don't keep
re-requesting as that is an infinate loop.